requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
streamlit==1.32.0
langchain==0.1.12
langchain-google-genai==1.0.1
crewai==0.22.4
sentence-transformers==2.5.1
faiss-cpu==1.8.0
spacy==3.7.4
pymupdf==1.23.26
python-docx==1.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
pydantic==2.6.3
python-dotenv==1.0.0
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
reportlab==4.1.0
python-multipart==0.0.9
